+++ resources/views/bootstrap-4/components/buttons/link-button.blade.php
@@ -10,8 +10,8 @@
         title="{!! $title !!}"
     @endif
     @if ($confirm !== null)
-        data-bs-confirm="{!! $confirm !!}"
-        data-bs-confirm-modal="confirm-modal-{!! $confirmId !!}"
+        data-buk-confirm="{!! $confirm !!}"
+        data-buk-confirm-modal="confirm-modal-{!! $confirmId !!}"
     @endif
 >
     @if ($startContent !== null)