+++ src/Components/Buttons/Actions/Enabled.php
@@ -17,7 +17,7 @@ class Enabled extends FormButton
 
         $this->text ??= Str::ucfirst(trans('status.enabled'));
 
-        $this->formId ??= 'disable-'.Str::random(32);
+        $this->formId = 'disable-'.($this->formId ?? Str::random(32));
     }
 
     public function viewName(): string