+++ src/Components/Buttons/SimpleButton.php
@@ -42,8 +42,8 @@ class SimpleButton extends BladeComponent
         $this->initAttributes();
         $this->onConstructing();
 
-        if ($this->confirm !== null && $this->confirmId === null) {
-            $this->confirmId = Str::random(32);
+        if ($this->confirm !== null) {
+            $this->confirmId = 'simple-button-'.($this->confirmId ?? Str::random(32));
         }
 
         $this->validBtnType(self::DEFAULT_BUTTON_TYPE);